Avoid circular dependency on QuantumControlBase
QuantumControlBase extends QuantumPropagators, but QuantumPropagators
does not depend on QuantumControlBase except for testing... Thus the
dependency should be only in test/Project.toml and docs/Project.toml,
but not in the main Project.toml.

This also makes it necessary to not use Pkg.test() for local
development: There's just no way to dev-install a local test-dependency
without that dependency being in the main Project.toml. So anything
that's only a test-dependency, but not a dependency of the package
itself, will break.  That's actually how the circular dependency got
into Project.toml in the first place.

On CI, we don't have to worry about this as much: the testing will
modify the main Project.toml, but we throw away the checkout after the
CI run finishes, so it works.

All of these problems are because Pkg.test() is has some buggy behavior
if there is a test/Manifest.toml (which would be the appropriate place
to record the local dev-installation of test dependencies):
https://github.com/JuliaLang/Pkg.jl/issues/1585
https://github.com/JuliaLang/Pkg.jl/issues/1233