Refactor to add underscores to method names
* `initprop` → `init_prop`
* `propstep!` → `prop_step!`
* `reinitprop!` → `reinit_prop!`
* `getcontrolderiv` → `get_control_deriv`
* `getcontrolderivs` → `get_control_derivs`
* `getcontrols` → `get_controls`

The old names were based on the Julia Style Guide recommendation
to name methods with "multiple words squashed together". I've come to
regard this as a very bad idea. To quote the JuMP style guide, "This
convention creates the potential for unnecessary bikeshedding and also
forces the user to recall the presence/absence of an underscore, e.g.,
"was that argument named basename or base_name?". For consistency,
always use underscores in variable names and function names to separate
words.