Add generalized `evaluate`, `substitute`
Rename `eval_controls` → `evaluate` and generalize such that `evaluate`
evaluates arbitrary time-dependent objects to static objects. The
`vals_dict` parameter for plugging in values for the controls is now an
optional keyword argument. Without it, the controls are simply evaluated
for the point in time specified by the positional arguments.

Also, rename `substitute_controls` → `substitute`. This can now
substitute a wide range of objects. One application this opens up is to
substitute controls with parametrized amplitudes.