Add `uniform_dt_tolerance` for Cheby propagator
Chebychev propagation requires a uniform time grid. We check that all the
`dt` in the time grid are the same. Sometimes, floating-point precision
issues can cause problems with this check. In that case, the option
`uniform_dt_tolerance` allows to set how close the `dt` in the time grid
have to be to each other for the time grid to be considered "uniform".