Allow Modules as propagation method, remove `:auto`
In preparation for allowing for ODE solvers via an optional dependency,
we now allow (and encourage) passing modules as a propagation `method`.

Also, the `method` now must be explicitly specified, `:auto` is no
longer an option. The heuristics for choosing a propagation methods are
brittle, and it's probably a good idea for people to think about which
method they should use anyway.

With optional dependencies (ODE) being added, automatically choosing a
method becomes even weirder.